Move `catalog::catalog` contents into parent.
