+++ datafusion/core/src/execution/context.rs
@@ -17,7 +17,7 @@
 
 //! [`SessionContext`] contains methods for registering data sources and executing queries
 use crate::{
-    catalog::catalog::{CatalogList, MemoryCatalogList},
+    catalog::{CatalogList, MemoryCatalogList},
     datasource::{
         datasource::TableProviderFactory,
         listing::{ListingOptions, ListingTable},
@@ -55,8 +55,8 @@ use arrow::{
 };
 
 use crate::catalog::{
-    catalog::{CatalogProvider, MemoryCatalogProvider},
     schema::{MemorySchemaProvider, SchemaProvider},
+    {CatalogProvider, MemoryCatalogProvider},
 };
 use crate::dataframe::DataFrame;
 use crate::datasource::{
